requirements.txt
asn1crypto==0.24.0
Babel==2.5.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
defusedxml==0.5.0
Django==2.0.5
django-dotenv==1.4.1
django-filter==1.1.0
django-formtools==2.1
django-otp==0.4.3
django-phonenumber-field==1.3.0
django-smart-selects==1.5.4
django-templated-mail==1.1.1
django-twilio==0.9.0
django-two-factor-auth==1.7.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.1.5
idna==2.6
nexmo==2.0.0
oauthlib==2.1.0
phonenumberslite==8.9.6
psycopg2==2.7.4
pycparser==2.18
PyJWT==1.6.3
PySocks==1.6.8
python3-openid==3.1.0
pytz==2018.3
qrcode==4.0.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
South==1.0.2
twilio==6.14.3
urllib3==1.22
